+++ src/facets/facets.facet-handler.service.js
@@ -98,7 +98,7 @@
             // Update state, and broadcast it to listening facets.
             function update(event, constraint) {
                 event.stopPropagation();
-                if (self.state.facets[constraint.id] !== constraint) {
+                if (!_.isEqual(self.state.facets[constraint.id], constraint)) {
                     self.state.facets[constraint.id] = constraint;
                     broadCastConstraints(EVENT_FACET_CONSTRAINTS);
                 }